chore: Replace TSLint with ESLint
This patch replaces TSLint with ESLint and adds ESLint plugins for handling TypeScript. TSLint is deprecated and in maintenance mode, while ESLint is still under active development.

The configured ESLint rules are a bit stricter than what we had enabled with TSLint, so the lint errors/warnings have been fixed by this PR.